refactor: update updateContact and deleteContact to use user.id

+++ backend/src/controllers/contactsController.js
@@ -32,6 +32,11 @@ const updateContact = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error('Contact not found');
     }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(401);
+        throw new Error('Not authorized to update this contact');
+
+    }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -45,6 +50,11 @@ const deleteContact = asyncHandler(async (req, res) => {
     if (!contact) {
         res.status(404);
         throw new Error('Contact not found');
+    }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(401);
+        throw new Error('Not authorized to update this contact');
+
     }
     await Contact.deleteOne({ _id: req.params.id });
     res.status(204).json();
